Refactor reporting logic to runner Fix aggregate report bulk loading oois
